Updated GameScene
