--- assets/game_manager/ChestModel.js
@@ -1,9 +1,0 @@
-class ChestModel {
-  constructor(x, y, gold, spawnerId) {
-    this.id = `${spawnerId}-${uuid.v4()}`;
-    this.spawnerId = spawnerId;
-    this.x = x;
-    this.y = y;
-    this.gold = gold;
-  }
-}
