--- assets/game_manager/utils.js
@@ -1,8 +1,0 @@
-const SpawnerType = {
-  MONSTER: 'MONSTER',
-  CHEST: 'CHEST',
-};
-
-function randomNumber(min, max) {
-  return Math.floor(Math.random() * max) + min;
-}
