+++ src/game_manager/utils.js
@@ -1,0 +1,8 @@
+export const SpawnerType = {
+  MONSTER: 'MONSTER',
+  CHEST: 'CHEST',
+};
+
+export function randomNumber(min, max) {
+  return Math.floor(Math.random() * max) + min;
+}
